MON-738: [Android] Public 2.1.0 Activation Demo name (#142)
* Initial commit - BlinkReceiptEngage - `com.blinkreceipt.engage `

* Initial Engage Activations integration

* 2.1.0-beta01

* Refactor demo app name "Activations" to "Activation"
