+++ blinkreceipt-demo/ActualActivation/build.gradle
