+++ blinkreceipt-demo/ActualActivation/proguard-rules.pro
