+++ blinkreceipt-demo/ActualActivation/src/main/java/com/actualplatform/android/activation/development/ui/RewardCurrencySection.kt
