+++ blinkreceipt-demo/ActualActivation/src/main/res/drawable/ic_arrow_back.xml
