+++ blinkreceipt-demo/BlinkReceiptEngage/build.gradle
@@ -1,0 +1,104 @@
+plugins {
+    id 'com.android.application'
+    id 'org.jetbrains.kotlin.plugin.compose'
+    id 'org.jetbrains.kotlin.plugin.serialization'
+}
+
+android {
+    namespace 'com.blinkreceipt.engage'
+    compileSdkVersion versions.compileSdk
+    buildToolsVersion versions.buildTools
+
+    defaultConfig {
+        applicationId "com.blinkreceipt.engage"
+        minSdkVersion versions.minSdk
+        targetSdkVersion versions.targetSdk
+
+        versionCode versions.versionMajor * 10000 + versions.versionMinor * 1000 + versions.versionPatch * 100
+        versionName "${versions.versionMajor}.${versions.versionMinor}.${versions.versionPatch}"
+
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
+    }
+
+    buildFeatures {
+        compose true
+    }
+
+    signingConfigs {
+        release {
+            storeFile file("../blink-receipt-demos.jsk")
+            storePassword "${keys.storePassword}"
+            keyAlias "${keys.keyAlias}"
+            keyPassword "${keys.keyPassword}"
+        }
+    }
+
+    buildTypes {
+        release {
+            minifyEnabled true
+            shrinkResources true
+            signingConfig signingConfigs.release
+            proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
+        }
+    }
+
+    compileOptions {
+        sourceCompatibility JavaVersion.VERSION_17
+        targetCompatibility JavaVersion.VERSION_17
+    }
+
+    packagingOptions {
+        resources {
+            excludes += [
+                    'META-INF/NOTICE.md',
+                    'META-INF/LICENSE.md',
+                    'META-INF/ASL2.0',
+                    'META-INF/AL2.0',
+                    'META-INF/LGPL2.1',
+                    'META-INF/licenses/ASM',
+                    'META-INF/DEPENDENCIES',
+            ]
+        }
+    }
+}
+
+kotlin {
+    jvmToolchain(17)
+}
+
+dependencies {
+    implementation "androidx.core:core-ktx:${androidx.core}"
+    implementation "androidx.activity:activity-compose:1.10.1"
+    implementation(platform("androidx.compose:compose-bom:2024.10.01"))
+    implementation "androidx.compose.material3:material3"
+    implementation "androidx.compose.ui:ui"
+    implementation "androidx.compose.ui:ui-tooling-preview"
+    debugImplementation "androidx.compose.ui:ui-tooling"
+    debugImplementation "androidx.compose.ui:ui-test-manifest"
+    androidTestImplementation "androidx.compose.ui:ui-test-junit4"
+
+    implementation(platform("com.microblink.blinkreceipt:blinkreceipt-bom:${versions.bom}"))
+    implementation("com.microblink.blinkreceipt:blinkreceipt-recognizer")
+    implementation("com.microblink.blinkreceipt:blinkreceipt-camera-ui")
+    implementation("com.microblink.blinkreceipt:blinkreceipt-logcat")
+
+    implementation(platform("com.actualplatform:engage-bom:${versions.engage}"))
+    implementation("com.actualplatform:engage")
+
+    implementation "androidx.navigation3:navigation3-runtime:${androidx.nav3Core}"
+    implementation "androidx.navigation3:navigation3-ui:${androidx.nav3Core}"
+    implementation "androidx.lifecycle:lifecycle-viewmodel-navigation3:${androidx.lifecycleViewModelNav3}"
+
+    implementation "com.squareup.okio:okio:${versions.okio}"
+    implementation "com.google.code.gson:gson:${versions.gson}"
+
+    testImplementation "junit:junit:${versions.junit}"
+
+    implementation "org.jetbrains.kotlinx:kotlinx-serialization-json:${versions.kotlinxSerialization}"
+
+    androidTestImplementation "androidx.test.ext:junit:${androidx.junit}"
+    androidTestImplementation "androidx.test.espresso:espresso-core:${versions.espresso}"
+
+    debugImplementation "androidx.compose.ui:ui-tooling"
+    debugImplementation "androidx.compose.ui:ui-test-manifest"
+}
