+++ blinkreceipt-demo/BlinkReceiptEngage/proguard-rules.pro
@@ -1,0 +1,3 @@
+# Add project specific ProGuard rules here.
+# You can control the set of applied configuration files using the
+# proguardFiles setting in build.gradle.
