+++ blinkreceipt-demo/BlinkReceiptEngage/src/main/java/com/blinkreceipt/engage/ui/ActivationsRoute.kt
@@ -1,0 +1,15 @@
+package com.blinkreceipt.engage.ui
+
+import androidx.navigation3.runtime.NavKey
+import kotlinx.serialization.Serializable
+
+internal sealed interface ActivationsRoute : NavKey {
+    @Serializable
+    data object Home : ActivationsRoute
+
+    @Serializable
+    data object OffersWall : ActivationsRoute
+
+    @Serializable
+    data object SettingsEditor : ActivationsRoute
+}
