+++ blinkreceipt-demo/BlinkReceiptEngage/src/main/java/com/blinkreceipt/engage/ui/EnvironmentSection.kt
@@ -1,0 +1,37 @@
+package com.blinkreceipt.engage.ui
+
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.RadioButton
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import com.blinkreceipt.engage.R
+
+@Composable
+internal fun EnvironmentSection(
+    environment: String,
+    onEnvironmentChange: (String) -> Unit,
+) {
+    Text(stringResource(R.string.engage_section_environment), style = MaterialTheme.typography.titleMedium)
+    Spacer(modifier = Modifier.height(8.dp))
+
+    listOf("Production", "Staging", "Development").forEach { env ->
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier.fillMaxWidth(),
+        ) {
+            RadioButton(
+                selected = environment == env,
+                onClick = { onEnvironmentChange(env) },
+            )
+            Text(env)
+        }
+    }
+}
