+++ blinkreceipt-demo/settings.gradle
@@ -5,7 +5,7 @@ include ':BlinkReceiptEarnings'
 include ':BlinkReceiptImap'
 include ':BlinkReceiptOutlook'
 include ':BlinkReceiptStaticScan'
-include ':ActualActivations'
+include ':ActualActivation'
 include ':BlinkReceiptCustomScan'
 include ':BlinkReceiptSurveys'
 include ':BlinkReceiptAccountLinking'
